requirements.txt
Flask>=2.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
black==24.2.0
openai==1.3.7
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pytest==7.4.3
requests==2.31.0 